Add components to display pokemon Rename fetch.js to pokemonFetch.js Add try, catch block in exported util function
Created the components PokemonContainer.jsx and Card.jsx to display the pokemon by using useEffect to synchronize the PokemonContainer component with the API. Renamed the util function to better describe what it does. Added error handling with the try, catch block in the util function, to catch possible errors and log them in the console.